Add files via upload
Includes a way to load model.
The variable, "result" is what the model should produce. 
"result[0]" is empty for some reason
"result[1]" contains predictions based on data given (not important to making a prediction on stocks in the future)
"result[2]" contains future predictions. (i.e. it will form predictions for the future)
Feel free to cut out the non-important stuff like graphing.
def split_data(stock, lookback) will need to be modified to fit the project.